Split files in Artboard
