+++ Shared/Sources/Artboard/ArtboardType.swift
@@ -1,0 +1,27 @@
+import Foundation
+
+public enum ArtboardType: String, Codable {
+    case element
+    case container
+    case frame
+    //    case artboard
+}
+
+public protocol ArtboardElement: Child, Equatable {
+    var id: UUID { get }
+    var label: String { get set }
+    var frame: CGRect { get set }
+    
+    var type: ArtboardType { get }
+}
+
+public protocol ArtboardContainer: ArtboardElement, Node {}
+
+/// Allows
+public protocol InstantiatableContainer {
+    init(
+        elements: [any ArtboardElement],
+        frame: CGRect,
+        label: String
+    )
+}
